+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 ## Version 0.0.0
 * ENH: move settings.py to settings.py.dist
+* FIX: django-registration 3.0 changed from 2.2 .. updated